dx(knip): Use knip instead of unimported
Unimported is deprecated and recommends switching to knip
