+++ packages/documentation/package.json
@@ -26,7 +26,6 @@
 		"@types/marked": "^6.0.0",
 		"@types/natural-sort": "^0.0.24",
 		"case-sensitive-paths-webpack-plugin": "^2.3.0",
-		"levenary": "^1.1.1",
 		"nuxt": "2.x",
 		"sass-loader": "10.x",
 		"ts-loader": "8.x",
@@ -46,7 +45,6 @@
 		"check:eslint": "yarn --cwd ../.. run eslint --max-warnings=0 --ignore-path=.gitignore ./packages/documentation",
 		"check:prettier": "yarn --cwd ../.. run prettier --check --ignore-path=.gitignore ./packages/documentation",
 		"check:stylelint": "stylelint './**/*.{css,scss,vue}'",
-		"check:unimported": "unimported --no-cache",
 		"fix:eslint": "yarn run check:eslint --fix",
 		"fix:prettier": "yarn run check:prettier --write",
 		"fix:stylelint": "yarn run check:stylelint --fix",
