+++ packages/fake-root/package.json
@@ -1,0 +1,12 @@
+{
+	"name": "fake-root",
+	"private": true,
+	"scripts": {
+		"check:eslint": "yarn --cwd ../.. run eslint --max-warnings=0 --ignore-pattern=packages .",
+		"check:knip": "yarn --cwd ../.. run knip",
+		"check:prettier": "yarn --cwd ../.. run prettier --check --ignore-path=.fake-prettierignore --ignore-path=.prettierignore .",
+		"fix:eslint": "yarn run check:eslint --fix",
+		"fix:prettier": "yarn run check:prettier --write"
+	},
+	"version": "0.0.0"
+}
