+++ packages/kotti-ui/package.json
@@ -83,7 +83,6 @@
 		"check:prettier": "yarn --cwd ../.. run prettier --check --ignore-path=.gitignore ./packages/kotti-ui",
 		"check:publint": "publint",
 		"check:stylelint": "stylelint 'source/**/*.{css,scss,vue}'",
-		"check:unimported": "unimported --no-cache",
 		"fix:eslint": "yarn run check:eslint --fix",
 		"fix:prettier": "yarn run check:prettier --write",
 		"fix:stylelint": "yarn run check:stylelint --fix",
