+++ packages/kotti-ui/source/constants.ts
@@ -1,4 +1,3 @@
-export const ISO8601 = 'YYYY-MM-DD' as const
 export const ISO8601_SECONDS = 'YYYY-MM-DD HH:mm:ss' as const
 
 // eslint-disable-next-line no-magic-numbers
