+++ packages/kotti-ui/source/kotti-table/mixins.d.ts
@@ -1,4 +1,3 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 declare const TableColumnsStateMixin: any
-declare const tableInjectedMixin: any
-export { TableColumnsStateMixin, tableInjectedMixin }
+export { TableColumnsStateMixin }
