+++ packages/vue-use-tippy/package.json
@@ -54,7 +54,6 @@
 		"check:eslint": "yarn --cwd ../.. run eslint --max-warnings=0 --ignore-path=.gitignore ./packages/vue-use-tippy",
 		"check:prettier": "yarn --cwd ../.. run prettier --check --ignore-path=.gitignore ./packages/vue-use-tippy",
 		"check:publint": "publint",
-		"check:unimported": "unimported --no-cache",
 		"fix:eslint": "yarn run check:eslint --fix",
 		"fix:prettier": "yarn run check:prettier --write"
 	},
