+++ packages/yoco/package.json
@@ -58,7 +58,6 @@
 		"check:prettier": "yarn --cwd ../.. run prettier --check --ignore-path=.gitignore ./packages/yoco",
 		"check:publint": "publint",
 		"check:stylelint": "stylelint style.css",
-		"check:unimported": "unimported --no-cache",
 		"fix:eslint": "yarn run check:eslint --fix",
 		"fix:prettier": "yarn run check:prettier --write"
 	},
